+++ slash/conf.py
@@ -17,7 +17,7 @@ config = Config({
         "truncate_console_lines": True // Doc("truncate long log lines on the console") // Cmdline(arg='--truncate-console-lines', metavar='yes/no'),
         "truncate_console_errors": False // Doc("If truncate_console_lines is set, also truncate long log lines, including and above the \"error\" level, on the console"),
         "root": None // Doc("Root directory for logs") // Cmdline(arg="-l", metavar="DIR"),
-        "subpath": "{context.session.id}/{context.test_id}/log" // Doc("Path to write logs to under the root"),
+        "subpath": "{context.session.id}/{context.test_id}/debug.log" // Doc("Path to write logs to under the root"),
         "session_subpath": "{context.session.id}/session.log",
         "last_session_symlink": None // Doc("If set, specifies a symlink path to the last session log file in each run"),
         "last_session_dir_symlink": None // Doc("If set, specifies a symlink path to the last session log directory"),
@@ -27,7 +27,8 @@ config = Config({
         "format": None // Doc("Format of the log line, as passed on to logbook. None will use the default format"),
         "console_format": None // Doc("Optional format to be used for console output. Defaults to the regular format"),
         "localtime": False // Doc("Use local time for logging. If False, will use UTC"),
-        "unittest_mode": False,
+        "unittest_mode": False // Doc("Used during unit testing. Emit all logs to stderr as well as the log files"),
+        "unified_session_log": False // Doc("Make the session log file contain all logs, including from tests"),
     },
     "run": {
         "default_sources": [] // Doc("Default tests to run assuming no other sources are given to the runner"),