+++ slash/core/scope_manager.py
@@ -16,10 +16,17 @@ class ScopeManager(object):
 
     def begin_test(self, test):
         test_module = test.__slash__.module_name
+        assert test_module
+
         if self._last_module is None:
             self._push_scope('session')
-            self._push_scope('module')
-        elif self._last_module != test_module:
+
+
+        if self._last_module != test_module:
+            if self._last_module is not None:
+                _logger.debug('Module scope has changed. Popping previous module scope')
+                self._pop_scope('module', in_failure=False, in_interruption=False)
+            assert self._scopes[-1] != 'module'
             self._push_scope('module')
         self._last_module = test_module
         self._push_scope('test')
@@ -32,19 +39,25 @@ class ScopeManager(object):
         kw = {'in_failure': exc_type is not None, 'in_interruption': exc_type is KeyboardInterrupt}
 
         self._pop_scope('test', **kw)
-        if next_test is None or next_test.__slash__.module_name != self._last_module:
-            self._pop_scope('module', **kw)
+
         if next_test is None:
+            _logger.debug('No next test. Popping scopes')
+            self._pop_scope('module', **kw)
             self._pop_scope('session', **kw)
 
+    def get_current_stack(self):
+        return self._scopes[:]
 
     def _push_scope(self, scope):
         self._scopes.append(scope)
+        _logger.debug('Pushed scope {0}', scope)
         self._session.fixture_store.push_scope(scope)
         self._session.cleanups.push_scope(scope)
 
     def _pop_scope(self, scope, **kw):
-        assert self._scopes.pop() == scope
+        popped = self._scopes.pop()
+        _logger.debug('Popped scope {0} (expected {1})', popped, scope)
+        assert popped == scope
         call_all_raise_first([self._session.cleanups.pop_scope, self._session.fixture_store.pop_scope],
                              scope, **kw)