+++ slash/frontend/slash_run.py
@@ -1,3 +1,4 @@
+import itertools
 import functools
 import os
 import sys
@@ -12,7 +13,7 @@ from ..exception_handling import handling_exceptions
 from ..resuming import (get_last_resumeable_session_id, get_tests_to_resume,
                         save_resume_state)
 from ..runner import run_tests
-from ..utils.interactive import start_interactive_shell
+from ..utils.interactive import generate_interactive_test
 
 _logger = logbook.Logger(__name__)
 
@@ -32,9 +33,9 @@ def slash_run(args, report_stream=None, resume=False, app_callback=None, test_so
                     collected = app.test_loader.get_runnables(to_resume, sort_key=test_sort_key)
                 else:
                     collected = _collect_tests(app, args, test_sort_key=test_sort_key)
-            with app.session.get_started_context():
                 if app.args.interactive:
-                    start_interactive_shell()
+                    collected = itertools.chain([generate_interactive_test()], collected)
+            with app.session.get_started_context():
                 run_tests(collected)
         except SlashException as e:
             logbook.error(str(e))