+++ tests/test_warnings.py
@@ -1,7 +1,7 @@
 import pytest
 import slash
 
-from .utils import run_tests_assert_success
+from .utils import run_tests_assert_success, without_pyc
 
 
 @pytest.mark.parametrize('reprify', [repr, str])
@@ -9,7 +9,7 @@ def test_str_repr(warning, reprify):
     assert 'this is a warning' in reprify(warning)
 
 def test_location(warning):
-    assert warning.details['filename'] == __file__
+    assert warning.details['filename'] == without_pyc(__file__)
 
 def test_to_dict(warning):
     assert isinstance(warning.to_dict(), dict)